+++ ESP32/rubiks-cube.cpp
@@ -23,12 +23,12 @@
 //     z /
 //
 
-void rotationX(Cube* cube, int columnNumber) {
+void rotationX(Cube* cube, uint8_t columnNumber) {
     if ( columnNumber > FACE_SIZE || columnNumber < 0) {
         return;
     }
-    int temp;
-    for (int i = 0; i < FACE_SIZE; ++i) {
+    uint8_t temp;
+    for (uint8_t i = 0; i < FACE_SIZE; ++i) {
         temp = *cube[0][i][columnNumber];
         *cube[0][i][columnNumber] = *cube[1][i][columnNumber];
         *cube[1][i][columnNumber] = *cube[5][FACE_SIZE-1-i][columnNumber];
@@ -37,12 +37,12 @@ void rotationX(Cube* cube, int columnNumber) {
     }
 }
 
-void rotationY(Cube* cube, int rowNumber) {
+void rotationY(Cube* cube, uint8_t rowNumber) {
     if ( rowNumber > FACE_SIZE || rowNumber < 0) {
         return;
     }
-    int temp;
-    for (int i = 0; i < FACE_SIZE; ++i) {
+    uint8_t temp;
+    for (uint8_t i = 0; i < FACE_SIZE; ++i) {
         temp = *cube[1][rowNumber][i];
         *cube[1][rowNumber][i] = *cube[2][rowNumber][i];
         *cube[2][rowNumber][i] = *cube[4][rowNumber][i];
@@ -51,12 +51,12 @@ void rotationY(Cube* cube, int rowNumber) {
     }
 }
 
-void rotationZ(Cube* cube, int profNumber) {
+void rotationZ(Cube* cube, uint8_t profNumber) {
     if ( profNumber > FACE_SIZE || profNumber < 0) {
         return;
     }
-    int temp;
-    for(int i = 0; i < FACE_SIZE; ++i){
+    uint8_t temp;
+    for(uint8_t i = 0; i < FACE_SIZE; ++i){
         temp = *cube[0][profNumber][i];
         *cube[0][profNumber][i] = *cube[3][FACE_SIZE-1-i][profNumber];
         *cube[3][FACE_SIZE-1-i][profNumber] = *cube[5][profNumber][i];
@@ -66,19 +66,19 @@ void rotationZ(Cube* cube, int profNumber) {
 }
 
 void setCubeToInitialPosition(Cube* cube) {
-    for (int i = 0; i < FACE_NUMBER; ++i) {
-        for (int j = 0; j < FACE_SIZE ; ++j) {
-            for (int k = 0; k < FACE_SIZE ; ++k) {
+    for (uint8_t i = 0; i < FACE_NUMBER; ++i) {
+        for (uint8_t j = 0; j < FACE_SIZE ; ++j) {
+            for (uint8_t k = 0; k < FACE_SIZE ; ++k) {
                 *cube[i][j][k] = i;
             }
         }
     }
 }
 
-void setCubeToColor(Cube* cube, int color) {
-    for (int i = 0; i < FACE_NUMBER; ++i) {
-        for (int j = 0; j < FACE_SIZE ; ++j) {
-            for (int k = 0; k < FACE_SIZE ; ++k) {
+void setCubeToColor(Cube* cube, uint8_t color) {
+    for (uint8_t i = 0; i < FACE_NUMBER; ++i) {
+        for (uint8_t j = 0; j < FACE_SIZE ; ++j) {
+            for (uint8_t k = 0; k < FACE_SIZE ; ++k) {
                 *cube[i][j][k] = color;
             }
         }
@@ -87,11 +87,11 @@ void setCubeToColor(Cube* cube, int color) {
 
 
 void displayCube(Cube* cube) {
-    for (int i = 0; i < FACE_NUMBER; ++i) {
+    for (uint8_t i = 0; i < FACE_NUMBER; ++i) {
         printf("\n face : %d \n", i);
-        for (int j = 0; j < FACE_SIZE ; ++j) {
+        for (uint8_t j = 0; j < FACE_SIZE ; ++j) {
             printf("\n");
-            for (int k = 0; k < FACE_SIZE ; ++k) {
+            for (uint8_t k = 0; k < FACE_SIZE ; ++k) {
                 printf(" %d ", *cube[i][j][k]);
             }
         }
