+++ ESP32/sys_conf.cpp
@@ -5,9 +5,9 @@
 void refreshCube(Cube *colorCube, Cube *ledMapping, Adafruit_NeoPixel *pixels){
 
     Serial.println("Refresh !");
-    for (int face = 0; face < FACE_NUMBER; ++face) {
-        for (int i = 0; i < FACE_SIZE; ++i) {
-            for (int j = 0; j < FACE_SIZE; ++j) {
+    for (uint8_t face = 0; face < FACE_NUMBER; ++face) {
+        for (uint8_t i = 0; i < FACE_SIZE; ++i) {
+            for (uint8_t j = 0; j < FACE_SIZE; ++j) {
                 if (*ledMapping[face][i][j] >= 0) {
                     pixels->setPixelColor(*ledMapping[face][i][j], getColorConfig(pixels, *colorCube[face][i][j]));
                 }
@@ -21,9 +21,9 @@ void refreshCube(Cube *colorCube, Cube *ledMapping, Adafruit_NeoPixel *pixels){
 void setLedMapping(Cube *ledMapping) {
     // init -1 everywhere
 
-    for (int face = 0; face < FACE_NUMBER; ++face) {
-        for (int i = 0; i < FACE_SIZE; ++i) {
-            for (int j = 0; j < FACE_SIZE; ++j) {
+    for (uint8_t face = 0; face < FACE_NUMBER; ++face) {
+        for (uint8_t i = 0; i < FACE_SIZE; ++i) {
+            for (uint8_t j = 0; j < FACE_SIZE; ++j) {
                 *ledMapping[face][i][j] = -1;
             }
         }
